split into multiple files & added verbose validation
