--- lib/mm_main.rb
@@ -1,2 +1,0 @@
-require_relative 'mm_code'
-
